fixed stupid change in schema

+++ RR-Calibration/src/test/java/com/scor/adjustment/service/adjustement/CalculAdjustementEEFReturnPeriodBandingTest.java
@@ -36,11 +36,11 @@ public class CalculAdjustementEEFReturnPeriodBandingTest {
     public void setUp() {
         log.info("Launch test for EEF Return Period Banding Adjustment");
         pltLossDataList = new ArrayList<PLTLossData>(){{
-            add(new PLTLossData(36,8443694,(double)41334,1,(double)3957801220L,(double)14240566));
-            add(new PLTLossData(68,8441785,(double)41542,1,(double)476993400L, 3455929.25));
-            add(new PLTLossData(74,8440073,(double)41512,1,(double)75025550L, 211684.75));
-            add(new PLTLossData(94,8440368,(double)41414,1,(double)292462560L, 112700.48));
-            add(new PLTLossData(117,8309116,(double)41292,1,(double)116508000000L, 12625397800L));
+            add(new PLTLossData(36,8443694,41334,1,(double)3957801220L,(double)14240566));
+            add(new PLTLossData(68,8441785,41542,1,(double)476993400L, 3455929.25));
+            add(new PLTLossData(74,8440073,41512,1,(double)75025550L, 211684.75));
+            add(new PLTLossData(94,8440368,41414,1,(double)292462560L, 112700.48));
+            add(new PLTLossData(117,8309116,41292,1,(double)116508000000L, 12625397800L));
         }};
         cap = true;
         periodConstante = 100000;
